Limit No Discharge/Battery Hold to 98% Soc to avoid PV from shutting down
The No Discharge option has an unintended side effect: When the SoC reaches 100% (and even when it later drops to 99%) an active No Discharge will make the inverter(s) to stop producing solar power. Even on the sunniest day, PV will drop to zero and all load will come from the grip, even if PV would produce more than the houseload.
In order to avoid this situation this commit limits the No Discharge option to below 98%. That way, even when the battery reached full due to excess PV production, the system will continue to produce full PV power.