Added payload.py from pyModbus to application
Brought the source file payload from the pyModbus project to the application. Done since payload.py is deprecated from pyModbus and causes a lot of warnings in the home assistant log.
To perform a rewrite of the logic more time is needed since it's a major rewrite since we should be able to easily split a multi fetched register (result)