X1-IES Inverter Temperature Incorrectly Scaled
On my X1-IES the inverter temperature was showing as 1/10th of the actual temperature.

It seems a new entry with 'scale=0.1' was added for the Gen 5 inverters which is doing this scaling. However this is incorrect for my inverter. I am guessing this scaling is only needed for the X3-IES variant, so adding a patch to split them up.