solis: add support for manual battery control
These new controls allow you to manually charge/discharge the battery
without having to configure timed charge/discharge. That is very useful
for automations.

This functionality requires an automated background process which keeps
resending the "manual control" command to the inverter or the inverter
will revert to regular operation in 5 minutes.