Add `battery_capacity` Sensor for SolaX AC Hybrid Gen5/6 Inverters
The Mode 8 Export-First remote control option requires a sensor called `battery_capacity`, however the hybrid Gen 5 and Gen 6 variants did not have this. Instead they have battery_1/2/total_capacity_charge sensors. This meant the remote control mode was not able to sense battery SoC and hence used grid import instead of battery when there was a deficit in PV.

To fix it, this commit adds a virtual `battery_capacity` sensor for the Gent 5/6 inverters which combines the three other data sources based on a best guess approach - essentially selecting based on which of the other sources is non-zero.

Fixes #1666.