Correct Mode 1 Export Limit Bound Check
The bound check was incorrectly adding on house load. This means the more the house uses, the less the inverter was allowed to output to the home. This is backwards.

Instead the more the house uses, the more the inverter should be allowed to output power so that it can continue to cover the house load, and still export to the grid up to the grid export limit.