Add Gain/Offset Calibration to Measured Power
The CT on my system has inaccuracies, both in terms of gain, but also a ~40W offset meaning the inverter thinks it is exporting more than it actually is.

There appears to be no way to calibrate this in the SolaX hybrid inverter itself which is somewhat annoying.

For the remote control modes, they require reasonably accurate `measured_power` values as they are basing their operation on "house load" which  is calculated from the measured power value. So here we add the ability to apply gain/offset correction to the measured power value in the form of a dynamic scale factor based on two new (disabled by default) entites.