Update plugin_growatt.py enhanced total PV total production logic
Added code to handle errors and unavailable states in the PV individual daily production when adding up to the total PV production.  This is to prevent PV total dropping to zero on an invalid read or HA reboot, which causes HA energy meters to double count the days PV production.